fix(ci): clear remaining Format & Clippy / Documentation / Miri errors
The previous CI-green commit only got as far as bcinr-logic's compile
errors before stopping, masking a large amount of pre-existing clippy
debt in bcinr-core, bcinr-powl, bcinr-cmca, and bcinr-logic itself that
`cargo clippy --workspace --all-targets -- -D warnings` had never
actually reached before. Also two rustdoc links and one Miri-hostile
test surfaced on the previous fix's own new code paths.

- bcinr-core/bcinr-powl/bcinr-cmca: remove unused imports/dead macros/
  dead code, convert needless_range_loop to iterator style, drop
  unnecessary same-type casts, fix manual assign-op/RangeInclusive/
  Clone-on-Copy patterns, allow clippy::too_many_arguments (with a
  justifying comment) on a handful of existing wide-parameter functions
  rather than redesigning their public signatures.
- bcinr-logic: same categories (identity_op no-ops in test bit-packing
  helpers, unnecessary casts, manual rotate/abs_diff/swap/sort_by_key/
  saturating-arithmetic/memcpy patterns, one inconsistent hex-literal
  casing, one byte-str literal) plus #[allow(clippy::result_unit_err)]
  on two public parse functions whose Result<_, ()> signature is
  intentional and out of scope to change here. Verified with
  `cargo test -p bcinr-logic --lib` (826 passed) since this crate's
  algorithms are formally verified.
- Add missing Cargo.toml `readme`/`keywords`/`categories`/`repository`
  metadata (house convention, matches bcinr-logic's Cargo.toml) to
  bcinr-pddl, bcinr-mfw-ir, bcinr-powl-receipt, bcinr-cmca, ggen.
- Delink two more broken rustdoc intra-doc links in bcinr-pddl's
  mfw/planner.rs pointing at a private item.
- Gate bcinr-cmca's compile_fail_tests (trybuild) with
  #![cfg(not(miri))]: trybuild globs test files via `std::fs::metadata`,
  which shells out to `statx`, unavailable under Miri's isolation
  sandbox — the same category as the earlier e2e_main.rs gate.
- Fix knock-on errors in tools/bcinr-contract-gate, bcinr's examples,
  and e2e tier3/tier4 tests that only became reachable once bcinr-logic
  and bcinr-cmca started compiling under -D warnings.

playground's ~200 pedantic-level clippy errors (from its own
`[lints.clippy] pedantic = "warn"` opt-in in playground/Cargo.toml) are
pre-existing on origin/main and untouched here — a separate, much
larger cleanup outside this PR's scope.

+++ crates/bcinr-cmca/src/fixed.rs
@@ -172,13 +172,13 @@ impl NonNegativeFixed {
         let q_shifted_46 = (q_u128 >> 46) as u64;
         let q = q_shifted_46 >> (32 - lz);
 
-        let rem = ((self.val as u64) << 16).wrapping_sub((q as u64).wrapping_mul(d as u64)) as i64;
+        let rem = ((self.val as u64) << 16).wrapping_sub(q.wrapping_mul(d as u64)) as i64;
 
         let is_lt = ((rem >> 63) & 1) as u64;
         let diff = rem.wrapping_sub(d as i64);
         let is_ge = (((!diff) >> 63) & 1) as u64;
 
-        let q_corrected = (q as u64).wrapping_add(is_ge).wrapping_sub(is_lt);
+        let q_corrected = q.wrapping_add(is_ge).wrapping_sub(is_lt);
 
         let overflow_1 = const_lt_u32(u32::MAX, (q_corrected >> 32) as u32).val;
         let overflow_2 = 0u32.wrapping_sub((q_corrected > u32::MAX as u64) as u32);
